gateway/pyproject.toml
[project]
name = "alfred-gateway"
version = "0.1.0"
description = "FastAPI gateway for Assistente Alfred"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "pydantic-settings>=2.6.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
